Modified midi note handling to accept and use MIDI_NOTE_ON with velocity 0 as synonymous with MIDI_NOTE_OFF.

+++ embedded/MidiKeys/debouncer.h
@@ -51,8 +51,8 @@ class debouncer {
 
  private:
   byte _value;
-  unsigned long _filter;
-  unsigned long _mask;
+  unsigned long long _filter;
+  unsigned long long _mask;
 };
 
 extern "C" {
